pre-release cleanup
