--- src/main/java/trade/wayruha/whitebit/Main.java
@@ -1,11 +1,0 @@
-package trade.wayruha.whitebit;
-
-public class Main {
-  public static void main(String[] args) {
-    System.out.println("Hello world!");
-  }
-
-  public static long getCurrentTime(){
-    return System.currentTimeMillis();
-  }
-}
