+++ src/main/java/trade/wayruha/whitebit/service/endpoint/PublicEndpoints.java
@@ -47,4 +47,18 @@ public interface PublicEndpoints {
     /** rate limit: 2000 requests/10sec */
     @GET("/api/v4/public/futures")
     Call<V1Response<List<FuturesMarket>>> getFuturesMarkets();
+
+    @GET("/api/v1/public/ticker")
+    Call<V1Response<MarketActivityV1.Ticker>> getAvailableTickerV1(@Query("market") Market symbol);
+
+    @GET("/api/v1/public/tickers")
+    Call<V1Response<Map<Market, MarketActivityV1>>> getAvailableTickersV1();
+
+    //V2
+    @GET("/api/v2/public/ticker")
+    Call<V1Response<List<MarketActivity>>> getAvailableTickersV2();
+
+    //V4
+    @GET("/api/v4/public/ticker")
+    Call<Map<Market, MarketActivityV4>> getAvailableTickersV4();
 }
